src/requirements.txt
asgiref==3.2.10
Django==3.1
django-environ==0.4.5
djangorestframework==3.11.1
Pillow==7.2.0
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
